==> JupyterNotebooks/1a ETL games.ipynb <==
# ETL DEL DATASET OUTPUT_STEAM_GAMES

#Se desarrolla la extracción, transformación y carga de output_steam_games.json

# IMPORTACIONES

import pandas as pd 
import json
import re

# FUNCIONES IMPORTANTES

# Para verificar tipos de datos y porcentajes
def verificar_tipo_dato(dataframe):
    diccionario = {'nombre_campo': [], 'tipo_dato': [], '%_no_nulos': [], '%_nulos': [], 'nulos': []}
    for columna in dataframe.columns:
        porcentaje_no_nulos = (dataframe[columna].count()/len(dataframe))*100
        diccionario['nombre_campo'].append(columna)
        diccionario['tipo_dato'].append(dataframe[columna].apply(type).unique())
        diccionario['%_no_nulos'].append(round(porcentaje_no_nulos,2))
        diccionario['%_nulos'].append(round(100-porcentaje_no_nulos,2))
        diccionario['nulos'].append(dataframe[columna].isnull().sum())
    
    df_info = pd.DataFrame(diccionario)
    return df_info





# Para ver la existencia de duplicados
def verificar_duplicados_columna(df, columna):
    lista_vacia = []
    filas_duplicadas = df[df.duplicated(subset=columna, keep=False)]
    for elemento in df[columna]:
        if len(filas_duplicadas) == len(lista_vacia):
            return 'No hay elementos duplicados'
        else:
            return filas_duplicadas.sort_values(by=columna, ascending=True)




# Para obtener año de lanzamiento
def obtener_año_lanzamiento(fecha):
    if pd.notna(fecha):
        if re.match(r'^\d{4}-\d{2}-\d{2}$', fecha):
            return fecha.split('-')[0]
    return 'Dato no disponible'




# Para reemplazar a float
def reemplazar_a_float(value):
    if pd.isna(value):
        return 0.0
    try:
        float_value = float(value)
        return float_value
    except:
        return 0.0

# EXTRACCION DE DATOS

# Se extraen los datos de los JSON y se les convierte en dataframe
archivo_games = 'C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\output_steam_games.json'

# Se lee el dataset por linea
filas_games = []
with open(archivo_games) as f:
    for line in f.readlines():
        data = json.loads(line)
        filas_games.append(data)
dfgames = pd.DataFrame(filas_games)
dfgames

# Inicialmente, se observan filas que presentan todos sus valores nulos, por lo que procedemos a eliminarlas. Si no tuvieramos esta primera visualización a mano, tendríamos que proceder a buscarlas de igual forma dentro del DataFrame
dfgames = dfgames.dropna(how='all').reset_index(drop=True)

dfgames.shape

#  Con funcion predeterminada se revisan los tipos de datos y la cantidad de nulos presentes
verificar_tipo_dato(dfgames)

# Luego de esta operación, aún quedan filas con valores nulos

# VERIFICACIÓN DE VALORES DUPLICADOS


# se constata si existen valores duplicados en este caso para la columna más importante, "id"
verificar_duplicados_columna(dfgames, 'id')

# se observan 2 id duplicados ("612880" y NaN). En primera instancia, podemos borrar el id 612880 (registro 14573) ya que son valores idénticos con el registro 13894. Por otro lado, hay que buscar a ver si los registros NaN corresponden al mismo juego. Haremos esto viendo el developer del registro 30961.

# buscamos el developer "Rocksteady Studios,Feral Interactive (Mac)" para ver a cual juego corresponde el valor NaN del registro 30961

dfgames[dfgames['developer'] == 'Rocksteady Studios,Feral Interactive (Mac)']

# de esta forma podemos ver que el id NaN corresponde por nombre de título al id 200260

# por lo tanto los índices a eliminar son:
indices_a_eliminar = [14573, 74, 30961]

dfgames = dfgames.drop(indices_a_eliminar)

# verificamos nuevamente duplicados
verificar_duplicados_columna(dfgames, 'id')

# COLUMNAS A TRANSFORMAR

# Se necesitan transformar las siguientes columnas:
# 'release_date': De esta columna se necesita extraer el año de lanzamiento. Crearemos una funcion para ello llamada "extraer_año_lanzamiento"
# 'price': Trabajaremos con esta columna, sin embargo los datos no están 'homogenizados' (como items "free to play" o promociones que están presentes como texto). Reemplazaremos estos valores por "0"
# 'publisher', 'app_name', 'title', 'developer', donde se observan valores nulos y se decide completarlos con valores "Dato no disponible"
# 'genres' la cual se presenta como una lista de valores, por lo que hay que desanidarlos 
# 'tags', 'specs', 'url' y 'reviews_url' que no se utilizaran ni para la API ni para el modelo de recomendacion

# COLUMNA 'release_date'

# creamos nueva columna 'anio_lanzamiento' y posteriormente eliminamos la columna 'release_date'
dfgames['anio_lanzamiento'] = dfgames['release_date'].apply(obtener_año_lanzamiento)

dfgames = dfgames.drop('release_date', axis=1)
dfgames.head()

dfgames['anio_lanzamiento'].unique()

# COLUMNA 'price'
dfgames['price'] = dfgames['price'].apply(reemplazar_a_float)
dfgames['price'].dtype

# COLUMNA 'publisher', 'app_name', 'title', 'developer'

columnas_a_completar = ['publisher', 'app_name', 'title', 'developer']

# se rellenan nulos
dfrelleno = dfgames[columnas_a_completar].fillna('Sin dato disponible')

# Se borran las columnas originales de dfgames y se concatena el relleno con todo el dataframe. 
dfgames = pd.concat([dfgames.drop(columnas_a_completar, axis=1), dfrelleno], axis=1)
dfgames.head()

# COLUMNA 'genres'

# esta columna presenta datos anidados en una lista, por lo que necesitamos desanidarlos y hacer una fila con cada genero. Para eso utilizaremos la funcion predeterminada "explode"

dfgames = dfgames.explode('genres')
dfgames = dfgames.dropna(subset='genres')
dfgames.head()

# COLUMNAS 'tags', 'specs', 'url' y 'reviews_url'

dfgames = dfgames.drop(['tags', 'specs', 'url', 'reviews_url'], axis=1)
dfgames.columns

# se verifican los datos y se observa si quedan nulos restantes
verificar_tipo_dato(dfgames)

#  GUARDADO DEL DATASET

#Se guarda el dataframe transformado como dfgames_limpio.
dfgames.to_csv(r"C:\Users\cquir\OneDrive\Escritorio\Data Science SH\Proyecto Individual 1\bases de datos\dfgames_limpio.csv", index=None)
