==> JupyterNotebooks/1b ETL reviews.ipynb <==
# ETL DEL DATASET AUSTRALIAN_USER_REVIEWS
 
#Se desarrolla la extracción, transformación y carga de australian_user_reviews.json

import pandas as pd 
import json
import re 
import ast 

# FUNCIONES IMPORTANTES

# Para verificar tipos de datos y porcentajes
def verificar_tipo_dato(dataframe):
    diccionario = {'nombre_campo': [], 'tipo_dato': [], '%_no_nulos': [], '%_nulos': [], 'nulos': []}
    for columna in dataframe.columns:
        porcentaje_no_nulos = (dataframe[columna].count()/len(dataframe))*100
        diccionario['nombre_campo'].append(columna)
        diccionario['tipo_dato'].append(dataframe[columna].apply(type).unique())
        diccionario['%_no_nulos'].append(round(porcentaje_no_nulos,2))
        diccionario['%_nulos'].append(round(100-porcentaje_no_nulos,2))
        diccionario['nulos'].append(dataframe[columna].isnull().sum())
    
    df_info = pd.DataFrame(diccionario)
    return df_info




# Para ver la existencia de duplicados
def verificar_duplicados_columna(df, columna):
    lista_vacia = []
    filas_duplicadas = df[df.duplicated(subset=columna, keep=False)]
    for elemento in df[columna]:
        if len(filas_duplicadas) == len(lista_vacia):
            return 'No hay elementos duplicados'
        else:
            return filas_duplicadas.sort_values(by=columna, ascending=True)
        



def convertir_fecha(fecha):
    match = re.search(r'(\w+\s\d{1,2},\s\d{4})', fecha)
    if match:
        fecha_str = match.group(1)
        try:
            fecha_dt = pd.to_datetime(fecha_str)
            return fecha_dt.strftime('%Y-%m-%d')
        except:
            return 'Fecha inválida'
    else:
        return 'Formato inválido'

# EXTRACCIÓN DE DATOS

# Se extraen los datos de los JSON y se les convierte en dataframe
rows=[]
with open("C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\australian_user_reviews.json", "r", encoding="UTF-8") as f:
    for line in f.readlines():
        rows.append(ast.literal_eval(line))
dfreviews = pd.DataFrame(rows)
dfreviews

# Se verifican tipos de datos y cantidad de nulos
verificar_tipo_dato(dfreviews)

# Se verifica si existen duplicados 
verificar_duplicados_columna(dfreviews, 'user_id')

# Se observan 623 filas duplicadas en 'user_id', pero hay que ver si para cada duplicado de user_id sus reviews son idénticas o difieren por ser más de un comentario hecho por el usuario. 
user_id = '05041129'
user_reviews = dfreviews[dfreviews['user_id'] == user_id]['reviews']
for review_list in user_reviews:
    for review in review_list:
        print(review['review'])
    print('-' * 40)

# En este caso los reviews son idénticos, por lo que se elimina la primera ocurrencia para cada uno de los user_id duplicados 
dfreviews = dfreviews.drop_duplicates(subset='user_id', keep='first')

verificar_duplicados_columna(dfreviews,'user_id')

# Se revisa la estructura de los datos presentes en la columna reviews 
dfreviews['reviews'][0]

# Se puede observar que es una lista de diccionarios hechos por cada videojuego diferente

# En este caso, debemos realizar 2 procedimientos a las diferentes variables de estos diccionarios:

# a. En primera instancia, transformar la columna general 'reviews', ya que está anidada y se busca generar una columna por cada diccionario para posteriormente crear un registro por cada diccionario 
# b. Transformar la columna 'reviews_posted', procesando la fecha y convirtiendola a un formato homogeneo y específico para todas 

# TRANSFORMACION 'reviews'
dfreviews2 = pd.json_normalize(dfreviews['reviews'])
dfreviews2.head()

# ya que transformamos solo la columna reviews, perdemos el 'user_id' y 'user_url'. Sin embargo, estos se encuentran en la misma posición original, por lo que se pueden concatenar con el dataframe anterior.

# Se añade 'user_id' y 'user_url' a las columnas previas 
dfreviews2 = pd.concat([dfreviews[['user_id', 'user_url']], dfreviews2], axis=1)
dfreviews2.head()

# dfreviews2.shape

#Tenemos los diccionarios con cada review ubicadas en columnas. Sin embargo, debemos generar un registro por cada diccionario, pasando de un formato ancho a uno largo mediante la función melt, manteniendo en cada caso el usuario que lo genesra.

dfreviews2 = pd.melt(dfreviews2, id_vars=['user_id', 'user_url'], value_vars=list(range(9)), value_name='reviews')
dfreviews2.head()

# se puede observar que con la funcion anterior, las columnas 'reviews' pasan a ser listas, pasando de un formato ancho a largo en el cual se tienen los registros de opiniones unicas asociadas a cada usuario. 
dfreviews2[dfreviews2['user_id']=='76561197970982479']

# Lo anterior genera valores None en las columnas 'reviews', que se proceden a eliminar
dfreviews2 = dfreviews2.dropna()

# Se verifica que cada 'user_id' tenga asociado solamente sus diccionarios correspondientes
dfreviews2[dfreviews2['user_id']=='76561197970982479']

# Ahora podemos convertir cada diccionario de la columna 'reviews' en columnas independientes con las claves como nombres de las columnas. Al desanidar los diccionarios de la columna 'reviews', cada clave o key se convierte en columna y cada value en fila. Esto se realiza mediante la función apply(pd.series) 
dfreviews = dfreviews2['reviews'].apply(pd.Series, dtype='object')
dfreviews = dfreviews.add_prefix('reviews_')
dfreviews

#Por lo anterior, nuevamente se pierde 'user_id' y 'user_url' por lo que se concatena de nuevo.
dfreviews = pd.concat([dfreviews2[['user_id', 'user_url']], dfreviews], axis=1)
dfreviews.head()

#Se observan valores faltantes en algunas columnas, probablemente deben tener un espacio ya que no son nulos como tal. 
dfreviews['reviews_last_edited'][0]

#Se reemplazan esos espacios por nulos.
dfreviews.replace('', None, inplace=True)
dfreviews.head()

# Verificamos como quedan nuestros datos luego de desanidar las columnas
verificar_tipo_dato(dfreviews)

# Se observa un gran % de valores nulos tanto en la columna 'reviews_funny' y 'reviews_last_edited', por lo que se decide eliminarlas. Por otro lado, la columna 'reviews_review' tiene un 5% de valores nulos, pero no se eliminan ya que se tomarán como un comentario neutral

dfreviews = dfreviews.drop(columns=['reviews_funny', 'reviews_last_edited'])
dfreviews.columns

# TRANSFORMACIÓN 'reviews_posted'

# necesitamos que la fecha sea en formato YYYY-MM-DD, sin embargo se encuentra como una cadena con otro formato. Procesaremos las fechas mediante expresiones regulares para extraer el año, mes y día
dfreviews['reviews_date'] = dfreviews['reviews_posted'].apply(convertir_fecha)
dfreviews['reviews_date']

dfreviews[dfreviews['reviews_date'] == 'Formato inválido']

#Hay 9771 registros que presentan un formato diferente al establecido ya que no presentan el año. Los registros no se podrán consultar desde la API, pero las otras columnas pueden aportar información relevante

# De esta forma se elimina la columna 'reviews_posted' ya que ya fue transformada y sustituida por la columna 'reviews_date'
dfreviews = dfreviews.drop(labels='reviews_posted', axis=1)

dfreviews.head()

verificar_tipo_dato(dfreviews)

# COLUMNA 'reviews_review'
# contiene 5% de valores nulos, por lo que se eliminan
dfreviews = dfreviews.dropna(subset=['reviews_review'])

#se verifican los nulos
verificar_tipo_dato(dfreviews)

#GUARDAR EL DATASET

dfreviews.to_csv (r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\dfreviews_limpio.csv", index = None)
