==> JupyterNotebooks/1c ETL items.ipynb <==
# ETL DEL DATASET AUSTRALIAN_USER_ITEMS
 
#Se desarrolla la extracción, transformación y carga de australian_user_items.json

import pandas as pd
import json
import re
import ast

# FUNCIONES IMPORTANTES

# Para verificar tipos de datos y porcentajes
def verificar_tipo_dato(dataframe):
    diccionario = {'nombre_campo': [], 'tipo_dato': [], '%_no_nulos': [], '%_nulos': [], 'nulos': []}
    for columna in dataframe.columns:
        porcentaje_no_nulos = (dataframe[columna].count()/len(dataframe))*100
        diccionario['nombre_campo'].append(columna)
        diccionario['tipo_dato'].append(dataframe[columna].apply(type).unique())
        diccionario['%_no_nulos'].append(round(porcentaje_no_nulos,2))
        diccionario['%_nulos'].append(round(100-porcentaje_no_nulos,2))
        diccionario['nulos'].append(dataframe[columna].isnull().sum())
    
    df_info = pd.DataFrame(diccionario)
    return df_info

# EXTRACCIÓN DE DATOS

# Se extraen los datos de los JSON y se les convierte en dataframe

rows=[]
with open("C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\australian_users_items.json", "r", encoding="utf-8") as f:
    for line in f.readlines():
        rows.append(ast.literal_eval(line))
dfitems = pd.DataFrame(rows)
dfitems

# Se verifican tipos de datos y cantidad de nulos
verificar_tipo_dato(dfitems)

# En cada registro de la columna 'items' hay una lista, así que se explora para cononcer su estructura
dfitems['items'][0]

# como corresponde a una lista de diccionarios, utilizamos la funcion json_normalize para poder convertir cada diccionario en una columna independiente. Sin embargo, esta funcion no es aplicada directamente al dataframe ya hecho, sino que se aplica a la lista de diccionarios creados inicialmente, ya que difiere en como se estructura el resultado. Al aplicar esta funcion sobre la lista de diccionarios creados inicialmente, cada diccionario para a ser un registro con sus claves como columnas y sus values como datos del registro. Por otro lado, si aplicasemos esta funcion a las columnas del dataframe, el reusltado sería otro dataframe con las claves de los diccionarios como columnas (los que en este caso pueden ser enormemente grandes, lo que no nos conviene)
dfitems_2 = pd.json_normalize(rows, record_path=['items'], meta=['steam_id','items_count','user_id', 'user_url'] )
dfitems_2 

# se verifican nuevamente los datos 
verificar_tipo_dato(dfitems_2)

# se averigua si existen duplicados
duplicados = dfitems_2.loc[dfitems_2.duplicated()]
duplicados

# Se observan 59104 duplicados, por lo que se procede a borrarlos
dfitems_2 = dfitems_2.drop_duplicates(keep='first')
dfitems_2

#La columna 'playtime_2weeks' se elimina porque la información del tiempo acumulado ya s eencuentra en 'playtime_forever'
dfitems_2 = dfitems_2.drop('playtime_2weeks', axis=1)

dfitems_2.columns

#GUARDAR australian_user_items
dfitems_2.to_csv (r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\dfitems_limpio.csv", index = None)
