==> JupyterNotebooks/2 datos para consulta.ipynb <==
# PREPARACIÓN DE DATOS PARA LAS CONSULTAS EN LA API 

import pandas as pd 
from textblob import TextBlob

# FUNCIONES IMPORTANTES

def verificar_tipo_datos(df):    
    mi_dict = {"nombre_campo": [], "tipo_datos": [], "no_nulos_%": [], "nulos_%": [], "nulos": []}
    for columna in df.columns:
        porcentaje_no_nulos = (df[columna].count() / len(df)) * 100
        mi_dict["nombre_campo"].append(columna)
        mi_dict["tipo_datos"].append(df[columna].apply(type).unique())
        mi_dict["no_nulos_%"].append(round(porcentaje_no_nulos, 2))
        mi_dict["nulos_%"].append(round(100-porcentaje_no_nulos, 2))
        mi_dict["nulos"].append(df[columna].isnull().sum())
    df_info = pd.DataFrame(mi_dict)
    return df_info




def analisis_de_sentimientos(review):
    if review is None:
        return 1
    analysis = TextBlob(review)
    polarity = analysis.sentiment.polarity
    if polarity < -0.2:
       return 0
    elif polarity > 0.2:
       return 2
    else:
       return 1




def ejemplos_review_por_sentimiento(reviews, sentiments):
    for sentiment_value in range(3):
        print(f'Para la categoría de análisis de sentimiento {sentiment_value} se tienen estos ejemplos de reviews:')
        
        '''sentiment_reviews = []
        for review, sentiment in zip(reviews, sentiments):
            sentiment_reviews.append(review)''' 
        #con comprension de listas la expresion anterior queda:
        sentiment_reviews = [review for review, sentiment in zip(reviews,sentiments) if sentiment == sentiment_value]
        for i, review in enumerate(sentiment_reviews[:3], start=1):
            print(f"Review {i}: {review}")
        print("\n")

# Extracción de datos

#Se extraen los datos de los diferentes datasets de games, reviews e items

dfgames = pd.read_csv("C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\dfgames_limpio.csv", encoding='utf-8')

dfreviews = pd.read_csv("C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\dfreviews_limpio.csv", encoding='utf-8')

dfitems = pd.read_csv("C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\dfitems_limpio.csv", encoding='utf-8')

# ANALISIS DE SENTIMIENTOS

#Se debe reemplazar la columna 'reviews_review' con una nueva llamada 'sentiment_analysis' donde se realice un análisis de sentimiento de los comentarios y que arrojará valores de la siguiente escala:

#0 si es malo,
#1 si es neutral o esta sin review
#2 si es positivo.

# Se realizará un análisis de sentimiento básico utilizando TextBlob que es una biblioteca de procesamiento de lenguaje natural (NLP). El objetivo de esta metodología es asignar un valor numérico a un texto, en este caso a los comentarios que los usuarios dejaron para un juego determinado, para representar si el sentimiento expresado en el texto es negativo, neutral o positivo.

#Esta metodología toma una revisión de texto como entrada, utiliza TextBlob para calcular la polaridad de sentimiento y luego clasifica la revisión como negativa, neutral o positiva en función de la polaridad calculada. En este caso, se consideraron las polaridades por defecto del modelo, el cuál utiliza umbrales -0.2 y 0.2, siendo polaridades negativas por debajo de -0.2, positivas por encima de 0.2 y neutrales entre medio de ambos.

dfreviews['sentiment_analysis'] = dfreviews['reviews_review'].apply(analisis_de_sentimientos)
dfreviews.head()

# Se revisan algunos ejemplos
ejemplos_review_por_sentimiento(dfreviews['reviews_review'], dfreviews['sentiment_analysis'])

# Se elimina la columna 'reviews_review' porque que ya no sirve
dfreviews = dfreviews.drop(columns=['reviews_review'])
dfreviews.columns

# Verificamos los datos
verificar_tipo_datos(dfreviews)

#Cantidad de dinero gastado por usuario y cantidad de items consumidos

#Se crea el dataframe 'df_gastos_items' que contiene por cada usuario el gasto en videosjuegos y la cantidad de items consumidos. Se crearán dataframe auxiliares y luego se unirán estos mismos

# Se extraen las columnas 'items_count', 'user_id' e 'item_id'
df_gastos_items = dfitems[['items_count', 'user_id', 'item_id']]
df_gastos_items

# Se hace un nuevo dataframe auxiliar, pero con los precios de los juegos
price_juegos = dfgames[['price', 'id']]

# Se eliminan duplicados
price_juegos = price_juegos.drop_duplicates(subset='id', keep='first')

# Se renombra 'id' por 'item_id' para unir luego
price_juegos = price_juegos.rename(columns={'id':'item_id'})
price_juegos

# Se unen ambos dataframes creados
df_gastos_items = df_gastos_items.merge(price_juegos, on='item_id', how='left')
df_gastos_items

# se verifica la información de los datos
verificar_tipo_datos(df_gastos_items)

# exploramos la columna 'price' ya que presenta valores nulos
df_gastos_items[df_gastos_items['price'].isnull()]

#Hay cerca de un 19% de juegos que no tienen precio. Asumimos que los que no tienen precio es porque son gratuitos y se los rellena con 0.0, ya que no contienen información.
df_relleno = df_gastos_items['price'].fillna(0.0)

## Se borra la columna original y se concatena la columna rellena con todo el dataframe
df_gastos_items = pd.concat([df_gastos_items.drop('price', axis=1), df_relleno], axis=1)
df_gastos_items

# Eliminamos la columna 'item_id' y agrupamos por usuario para calcular la cantidad gaastada
df_gastos_items = df_gastos_items.drop('item_id', axis=1)

# Se agrupa por usuario, sumando el precio total de items consumidos
df_gastos_items_group = df_gastos_items.groupby('user_id')['price'].sum().reset_index()
df_gastos_items_group 

#Para volver a agregar los items consumidos por cada usuario se hace un dataframe auxiliar 'df_count_items' conteniendo la cantidad de items por usuario.
df_count_items = df_gastos_items[['items_count', 'user_id']]

# se eliminan duplicados
df_count_items = df_count_items.drop_duplicates(subset='user_id', keep='first')
df_count_items

# Esto nos dará un dataframe con un unico registro para cada usuario de cuantos juegos ha comprado

#Finalmente, se juntan los dos dataframe auxiliares 'df_gastos_items_group' que contiene el gasto de cada usuario con 'df_count_items' que contiene la cantidad de items consumidos por cada usuario.
df_gastos_items = df_count_items.merge(df_gastos_items_group, on='user_id', how='right')
df_gastos_items

#Playtime_forever por usuario

#Generamos un dataframe que contenga por género de videojuego, el tiempo jugado por cada usuario con su id y url del perfil. Para esto generamos dataframes auxiliares

#En primer lugar, se extrae de df_items las columnas 'playtime_forever', 'user_id' y 'item_id' y se guarda en un nuevo dataframe.
playtime_forever_usuario_item = dfitems[['playtime_forever', 'user_id', 'item_id']]
playtime_forever_usuario_item

#Del dfgames extraemos 'id' del item y el género.
genre_item = dfgames[['genres', 'id']]

# Renombramos la columna 'id' y se une con el dataframe previo "playtime_forever_usuario_item"
genre_item = genre_item.rename(columns={'id':'item_id'})
genre_item

#Ahora se unen las dos tablas anteriores para obtener los generos de todos los usuarios con sus id y url.
df_playtime_genre = playtime_forever_usuario_item.merge(genre_item, on='item_id')
df_playtime_genre

#Agrupamos por género y usuario y se suma el tiempo de juego de para cada caso. Transformamos la columna 'playtime_forever' a horas, ya que se asume en minutos.

agg_genero = df_playtime_genre.groupby(['genres', 'user_id'])['playtime_forever'].sum().reset_index()

agg_genero['playtime_horas'] = agg_genero['playtime_forever']/60

# Borramos la columna playtime_forever, ya que la transformamos a horas
agg_genero = agg_genero.drop('playtime_forever', axis=1)
agg_genero

# Hacemos un dataframe auxiliar ya que buscamos almacenar el url del usuario y luego se unen con el dataframe auxiliar previo. Así generamos el dataframe 'df_playtime_forever' que contiene por género las horas de juego por usuario y su url.
user_url_usuario = dfitems[['user_url', 'user_id']]

# Se eliminan duplicados para tener una sola vez los 'user_id' con su url
user_url_usuario = user_url_usuario.drop_duplicates(subset='user_id', keep='first')

# Se une con el dataframe agregado de género
df_playtime_forever = agg_genero.merge(user_url_usuario, on='user_id', how='left')
df_playtime_forever

#Cantidad de items y contenido free por desarrollador

#Se hace un dataframe en el que para cada desarrollador contiene los items que desarrolla, el año de lanzamiento y su precio.
price_anio_developer = dfgames[['price', 'anio_lanzamiento', 'developer', 'id']]

# Se renombra 'id' para unirlo con el dataframe creado previamente
price_anio_developer = price_anio_developer.rename(columns={'id':'item_id'})

# se eliminan duplicados
df_items_developer = price_anio_developer.drop_duplicates()
df_items_developer

#Año de lanzamiento ('release_anio') en dfreviews

#A dfreviews se agregan los años de lanzamiento de cada juego. Primero se extraen las columnas de id y el año de lanzamiento del mismo, se borran los duplicados y se unen con dfreviews.
anio_lanzamiento_item = dfgames[['id', 'anio_lanzamiento']]

# Se renombra la columna 'id' para poder unirla con el dataframe previo
anio_lanzamiento_item = anio_lanzamiento_item.rename(columns={'id':'reviews_item_id'})

# se eliminan los duplicados
anio_lanzamiento_item = anio_lanzamiento_item.drop_duplicates()
anio_lanzamiento_item

dfreviews = dfreviews.merge(anio_lanzamiento_item, on='reviews_item_id')
dfreviews

# Desarrolladores_recomendados

#Se crea el dataframe 'desarrolladores_recomendados' que contiene por cada desarrollador el 'item_id' de cada juego, tanto como 'reviews_recommend' y 'sentiment_analysis'. Se crearán dataframe auxiliares y luego se unirán estos mismos

# Primero, se extraen de df_reviews las columnas necesarias para realizar un merge

recommend_sentiment = dfreviews[['user_id','reviews_item_id', 'reviews_recommend', 'sentiment_analysis']]

# Se renombra la columna 'reviews_item_id'
recommend_sentiment = recommend_sentiment.rename(columns={'reviews_item_id':'item_id'})
recommend_sentiment 

# merge para 'recommend_sentiment'

# Utilizamos el dataframe creado anteriormente 'df_items_developer' que contiene el item_id, año de lanzamiento y developer
dfdesarrolladores_recomendados = pd.merge(recommend_sentiment, df_items_developer, on='item_id', how='left')
dfdesarrolladores_recomendados

#Guardar dataframes a CSV

dfgames.to_csv (r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\dfgames_unido.csv", index = None, encoding='utf-8')

dfreviews.to_csv (r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\dfreviews_unido.csv", index = None, encoding='utf-8')

dfitems.to_csv (r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\dfitems_unido.csv", index = None, encoding='utf-8')

dfdesarrolladores_recomendados.to_csv (r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\dfdesarrolladores_recomendados_unido.csv", index = None, encoding='utf-8')

df_gastos_items.to_csv (r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\df_gastos_items_unido.csv", index = None, encoding='utf-8')
 
df_playtime_forever.to_csv (r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\df_playtime_forever_unido.csv", index = None, encoding='utf-8')

df_items_developer.to_csv (r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\df_items_developer_unido.csv", index = None, encoding='utf-8')

# Guardar dataframes a parquet
dfgames.to_parquet(r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\dfgames_unido.parquet", index = None)

dfreviews.to_parquet(r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\dfreviews_unido.parquet", index = None)

dfitems.to_parquet(r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\dfitems_unido.parquet", index = None)

dfdesarrolladores_recomendados.to_parquet(r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\dfdesarrolladores_recomendados_unido.parquet", index = None)

df_gastos_items.to_parquet(r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\df_gastos_items_unido.parquet", index = None)

df_playtime_forever.to_parquet(r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\df_playtime_forever_unido.parquet", index = None)

df_items_developer.to_parquet(r"C:\\Users\\cquir\\OneDrive\\Escritorio\\Data Science SH\\Proyecto Individual 1\\bases de datos\\df_items_developer_unido.parquet", index = None)
